refactor: reorganiza estrutura de pastas e ajusta imports para padrão profissional
